support overriding data bag settings with ['pxe_dust']['default'] attributes
The data bag code was actually greatly simplified by merging the defaults with
the node attributes, then merging this again with the image hash. All the '||'
everywhere were removed to make the code simpler. Regexes were used for matching
the Chef installers to the right versions, this was synced with the install.sh
and http://www.opscode.com/chef/install/ page for reference. A few foodcritic
and style cleanups went in as well.